requirements.txt
Flask~=2.1.2
onnxruntime~=1.11.1
numpy~=1.21.6
Pillow~=9.1.1